pyproject.toml
[project]
name = "FedoraCoffee-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Enzo Lazzarin Dal'osto", email = "wecredsbr@gmail.com"},
]
dependencies = [
    "django>=5.0.3",
    "djangorestframework>=3.15.2",
    "setuptools>=72.1.0",
    "Pillow>=10.4.0",
    "python-magic>=0.4.27; sys_platform == \"linux\"",
    "python-magic-bin>=0.4.14; sys_platform == \"win32, darwin\"",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
